Add Makefile target for build and use -mod=mod to avoid requiring vendor directory when running targets
