+++ Makefile
@@ -1,3 +1,8 @@
 .PHONY: test
 test:
-	go test ./...
+	go test -mod=mod ./...
+
+
+.PHONY: build
+build:
+	go build -mod=mod ./...
