+++ identity/identity_suite_test.go
@@ -1,436 +1,13 @@
 package identity_test
 
 import (
-	"context"
-	"encoding/base64"
-	"net/http"
-	"net/http/httptest"
 	"testing"
 
-	"github.com/redhatinsights/platform-go-middlewares/v2/identity"
-
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
-
-// example header based on x-rh-identity seen in stage env
-var exampleHeader = `
-{
-  "entitlements": {
-    "insights": {
-      "is_entitled": true,
-      "is_trial": false
-    },
-    "cost_management": {
-      "is_entitled": true,
-      "is_trial": false
-    }
-  },
-  "identity": {
-    "internal": {
-      "cross_access": false,
-      "auth_time": 0,
-      "org_id": "1979710"
-    },
-    "user": {
-      "is_active": true,
-      "locale": "en_US",
-      "is_org_admin": true,
-      "username": "Test",
-      "email": "test@test.com",
-      "first_name": "Test",
-      "user_id": "55555555",
-      "last_name": "User",
-      "is_internal": true
-    },
-    "auth_type": "jwt-auth",
-    "org_id": "1979710",
-    "account_number": "540155",
-    "type": "User"
-  }
-}
-`
-
-var serviceAccountIdentity = `
-{
-  "identity": {
-    "account_number": "540155",
-    "auth_type": "jwt-auth",
-    "internal": {},
-    "org_id": "1979710",
-    "service_account": {
-      "client_id": "0000",
-      "username": "jdoe",
-      "user_id": "5d16465b-c0be-4cf6-a26f-084ebbc5e67d"
-    },
-    "type": "ServiceAccount"
-  }
-}
-`
-
-var validJson = [...]string{
-	`{ "identity": {"account_number": "540155", "auth_type": "jwt-auth", "org_id": "1979710", "type": "User", "internal": {"org_id": "1979710"} } }`,
-	`{ "identity": {"account_number": "540155", "auth_type": "cert-auth", "org_id": "1979710", "type": "Associate", "internal": {"org_id": "1979710"} } }`,
-	`{ "identity": {"account_number": "540155", "auth_type": "basic-auth", "type": "Associate", "internal": {"org_id": "1979710"} } }`,
-	`{ "identity": {"account_number": "540155", "auth_type": "cert-auth", "org_id": "1979710", "type": "Associate", "internal": {} } }`,
-	exampleHeader,
-	serviceAccountIdentity,
-}
-
-func GetTestHandler(allowPass bool) http.HandlerFunc {
-	fn := func(rw http.ResponseWriter, req *http.Request) {
-		if !allowPass {
-			panic("test entered test handler, this should not happen")
-		}
-	}
-
-	return http.HandlerFunc(fn)
-}
-
-func boilerWithCustomHandler(req *http.Request, expectedStatusCode int, expectedBody string, handlerFunc http.HandlerFunc) {
-	rr := httptest.NewRecorder()
-
-	handler := identity.EnforceIdentity(handlerFunc)
-	handler.ServeHTTP(rr, req)
-
-	Expect(rr.Body.String()).To(Equal(expectedBody))
-	Expect(rr.Code).To(Equal(expectedStatusCode))
-	Expect(rr.Body.String()).To(Equal(expectedBody))
-}
-
-func boiler(req *http.Request, expectedStatusCode int, expectedBody string) {
-	rr := httptest.NewRecorder()
-	handler := identity.EnforceIdentity(GetTestHandler(expectedStatusCode == 200))
-	handler.ServeHTTP(rr, req)
-
-	Expect(rr.Code).To(Equal(expectedStatusCode))
-	Expect(rr.Body.String()).To(Equal(expectedBody))
-}
-
-func noopLogger(ctx context.Context, rawId, msg string) {
-}
-
-func boilerWithLogger(req *http.Request, expectedStatusCode int, expectedBody string) {
-	rr := httptest.NewRecorder()
-
-	handler := identity.EnforceIdentityWithLogger(noopLogger)
-	handler(GetTestHandler(expectedStatusCode == 200)).ServeHTTP(rr, req)
-
-	Expect(rr.Code).To(Equal(expectedStatusCode))
-	Expect(rr.Body.String()).To(Equal(expectedBody))
-}
-
-func getBase64(data string) string {
-	return base64.StdEncoding.EncodeToString([]byte(data))
-}
 
 func TestIdentity(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "Identity Suite")
 }
-
-var _ = Describe("Identity", func() {
-	var req *http.Request
-
-	BeforeEach(func() {
-		r, err := http.NewRequest("GET", "/api/entitlements/v1/services/", nil)
-		if err != nil {
-			panic("Test error unable to get a NewRequest")
-		}
-		req = r
-	})
-
-	Context("WithIdentity a valid x-rh-id header", func() {
-		It("should 200 and set raw identity on the context", func() {
-			for _, jsonIdentity := range validJson {
-				req.Header.Set("x-rh-identity", getBase64(jsonIdentity))
-
-				boilerWithCustomHandler(req, 200, "", func() http.HandlerFunc {
-					fn := func(rw http.ResponseWriter, nreq *http.Request) {
-						raw := identity.GetRawIdentity(nreq.Context())
-						id, err1 := identity.DecodeAndCheckIdentity(raw)
-						Expect(err1).To(BeNil())
-						Expect(id.Identity.OrgID).To(Equal("1979710"))
-						Expect(id.Identity.AccountNumber).To(Equal("540155"))
-					}
-					return http.HandlerFunc(fn)
-				}())
-			}
-		})
-
-		It("should 200 and set the org_id on the context", func() {
-			for _, jsonIdentity := range validJson {
-				req.Header.Set("x-rh-identity", getBase64(jsonIdentity))
-
-				boilerWithCustomHandler(req, 200, "", func() http.HandlerFunc {
-					fn := func(rw http.ResponseWriter, nreq *http.Request) {
-						id := identity.GetIdentity(nreq.Context())
-						Expect(id.Identity.OrgID).To(Equal("1979710"))
-						Expect(id.Identity.AccountNumber).To(Equal("540155"))
-					}
-					return http.HandlerFunc(fn)
-				}())
-			}
-		})
-
-		It("should 200 and set the entitlements on the context", func() {
-			req.Header.Set("x-rh-identity", getBase64(exampleHeader))
-
-			boilerWithCustomHandler(req, 200, "", func() http.HandlerFunc {
-				fn := func(rw http.ResponseWriter, nreq *http.Request) {
-					id := identity.GetIdentity(nreq.Context())
-					Expect(id.Entitlements["insights"].IsEntitled).To(BeTrue())
-					Expect(id.Entitlements["cost_management"].IsEntitled).To(BeTrue())
-				}
-				return http.HandlerFunc(fn)
-			}())
-		})
-
-		It("should be able to return the header again if headers are requested", func() {
-			for _, jsonIdentity := range validJson {
-				req.Header.Set("x-rh-identity", getBase64(jsonIdentity))
-
-				boilerWithCustomHandler(req, 200, "", func() http.HandlerFunc {
-					fn := func(rw http.ResponseWriter, nreq *http.Request) {
-						h := identity.EncodeIdentity(nreq.Context())
-						Expect(h).ToNot(BeEmpty())
-					}
-					return http.HandlerFunc(fn)
-				}())
-			}
-		})
-	})
-
-	Context("WithIdentity a valid x-rh-id header and logger", func() {
-		It("should throw a 400 with a descriptive message", func() {
-			boilerWithLogger(req, 400, "Bad Request: missing x-rh-identity header\n")
-		})
-	})
-
-	Context("WithIdentity a missing x-rh-id header", func() {
-		It("should throw a 400 with a descriptive message", func() {
-			boiler(req, 400, "Bad Request: missing x-rh-identity header\n")
-		})
-
-		It("should return empty string if headers are requested", func() {
-			boilerWithCustomHandler(req, 400, "Bad Request: missing x-rh-identity header\n", func() http.HandlerFunc {
-				fn := func(rw http.ResponseWriter, nreq *http.Request) {
-					h := identity.EncodeIdentity(nreq.Context())
-					Expect(h).To(BeEmpty())
-				}
-				return http.HandlerFunc(fn)
-			}())
-		})
-	})
-
-	Context("WithIdentity invalid b64 data in the x-rh-id header", func() {
-		It("should throw a 400 with a descriptive message", func() {
-			for _, jsonIdentity := range validJson {
-				req.Header.Set("x-rh-identity", "="+getBase64(jsonIdentity))
-				boiler(req, 400, "Bad Request: unable to b64 decode x-rh-identity header\n")
-			}
-		})
-	})
-
-	Context("WithIdentity invalid json data (valid b64) in the x-rh-id header", func() {
-		It("should throw a 400 with a descriptive message", func() {
-			for _, jsonIdentity := range validJson {
-				req.Header.Set("x-rh-identity", getBase64(jsonIdentity+"}"))
-				boiler(req, 400, "Bad Request: x-rh-identity header does not contain valid JSON: invalid character '}' after top-level value\n")
-			}
-		})
-
-		It("should return empty string if headers are requested", func() {
-			for _, jsonIdentity := range validJson {
-				req.Header.Set("x-rh-identity", getBase64(jsonIdentity+"}"))
-				boilerWithCustomHandler(req, 400, "Bad Request: x-rh-identity header does not contain valid JSON: invalid character '}' after top-level value\n", func() http.HandlerFunc {
-					fn := func(rw http.ResponseWriter, nreq *http.Request) {
-						h := identity.EncodeIdentity(nreq.Context())
-						Expect(h).To(BeEmpty())
-					}
-					return http.HandlerFunc(fn)
-				}())
-			}
-		})
-	})
-
-	Context("WithIdentity missing account_number in the x-rh-id header", func() {
-		It("should 200", func() {
-			req.Header.Set("x-rh-identity", getBase64(`{ "identity": {"org_id": "1979710", "auth_type": "basic-auth", "type": "Associate", "internal": {"org_id": "1979710"} } }`))
-
-			boilerWithCustomHandler(req, 200, "", func() http.HandlerFunc {
-				fn := func(rw http.ResponseWriter, nreq *http.Request) {
-					id := identity.GetIdentity(nreq.Context())
-					Expect(id.Identity.OrgID).To(Equal("1979710"))
-					Expect(id.Identity.Internal.OrgID).To(Equal("1979710"))
-					Expect(id.Identity.AccountNumber).To(Equal(""))
-				}
-				return http.HandlerFunc(fn)
-			}())
-		})
-	})
-
-	Context("WithIdentity a valid x-rh-id header", func() {
-		It("should 200 and set the type to associate", func() {
-			req.Header.Set("x-rh-identity", getBase64(`{ "identity": {"type": "Associate"} }`))
-
-			boilerWithCustomHandler(req, 200, "", func() http.HandlerFunc {
-				fn := func(rw http.ResponseWriter, nreq *http.Request) {
-					id := identity.GetIdentity(nreq.Context())
-					Expect(id.Identity.Type).To(Equal("Associate"))
-				}
-				return http.HandlerFunc(fn)
-			}())
-		})
-
-		It("should 200 and set the type to X509", func() {
-			req.Header.Set("x-rh-identity", getBase64(`{"identity": {"type": "X509"}}`))
-
-			boilerWithCustomHandler(req, 200, "", func() http.HandlerFunc {
-				fn := func(rw http.ResponseWriter, nreq *http.Request) {
-					id := identity.GetIdentity(nreq.Context())
-					Expect(id.Identity.Type).To(Equal("X509"))
-				}
-				return http.HandlerFunc(fn)
-			}())
-		})
-
-		It("should 200 and set the correct values for a ServiceAccount identity", func() {
-			req.Header.Set("x-rh-identity", getBase64(serviceAccountIdentity))
-
-			boilerWithCustomHandler(req, 200, "", func() http.HandlerFunc {
-				fn := func(rw http.ResponseWriter, nreq *http.Request) {
-					id := identity.Get(nreq.Context())
-					Expect(id.Identity.Type).To(Equal("ServiceAccount"))
-					Expect(id.Identity.ServiceAccount.Username).To(Equal("jdoe"))
-					Expect(id.Identity.ServiceAccount.ClientId).To(Equal("0000"))
-					Expect(id.Identity.ServiceAccount.UserId).To(Equal("5d16465b-c0be-4cf6-a26f-084ebbc5e67d"))
-				}
-				return http.HandlerFunc(fn)
-			}())
-		})
-	})
-
-	Context("WithIdentity missing org_id in the x-rh-id header", func() {
-		It("should throw a 400 with a descriptive message", func() {
-			var missingOrgIDJson = [...]string{
-				`{ "identity": {"account_number": "540155", "type": "User", "internal": {} } }`,
-			}
-
-			for _, jsonIdentity := range missingOrgIDJson {
-				req.Header.Set("x-rh-identity", getBase64(jsonIdentity))
-				boiler(req, 400, "Bad Request: x-rh-identity header has an invalid or missing org_id\n")
-			}
-		})
-	})
-
-	Context("WithIdentity missing type in the x-rh-id header", func() {
-		It("should throw a 400 with a descriptive message", func() {
-			req.Header.Set("x-rh-identity", getBase64(`{"identity":{"account_number":"540155","type":"", "org_id":"1979710", "internal": {"org_id": "1979710"}}}`))
-			boiler(req, 400, "Bad Request: x-rh-identity header is missing type\n")
-		})
-	})
-})
-
-var _ = Describe("IdentityTokenEncapsulator", func() {
-	Context("XRHID implementation", func() {
-		It("should return the organization ID from the identity", func() {
-			xrhid := identity.XRHID{
-				Identity: identity.Identity{
-					OrgID: "test-org-123",
-				},
-			}
-
-			orgID := xrhid.GetOrganizationID()
-			Expect(orgID).To(Equal("test-org-123"))
-		})
-
-		It("should return empty string when org ID is not set", func() {
-			xrhid := identity.XRHID{
-				Identity: identity.Identity{},
-			}
-
-			orgID := xrhid.GetOrganizationID()
-			Expect(orgID).To(Equal(""))
-		})
-	})
-
-	Context("NewXRHIDEncapsulator constructor", func() {
-		It("should create a valid XRHID encapsulator from valid JSON", func() {
-			validJSON := `{
-				"identity": {
-					"org_id": "test-org-456",
-					"type": "User",
-					"account_number": "123456"
-				}
-			}`
-
-			encapsulator, err := identity.NewXRHIDFromHeader([]byte(validJSON))
-
-			Expect(err).To(BeNil())
-			Expect(encapsulator).ToNot(BeNil())
-			Expect(encapsulator.GetOrganizationID()).To(Equal("test-org-456"))
-		})
-
-		It("should return error for empty input", func() {
-			encapsulator, err := identity.NewXRHIDFromHeader([]byte{})
-
-			Expect(err).To(Equal(identity.ErrMissingIdentity))
-			Expect(encapsulator).To(BeEquivalentTo(identity.XRHID{}))
-		})
-
-		It("should return error for nil input", func() {
-			encapsulator, err := identity.NewXRHIDFromHeader(nil)
-
-			Expect(err).To(Equal(identity.ErrMissingIdentity))
-			Expect(encapsulator).To(BeEquivalentTo(identity.XRHID{}))
-		})
-
-		It("should return error for invalid JSON", func() {
-			invalidJSON := `{"invalid": json}`
-
-			encapsulator, err := identity.NewXRHIDFromHeader([]byte(invalidJSON))
-
-			Expect(err).ToNot(BeNil())
-			Expect(encapsulator).To(BeEquivalentTo(identity.XRHID{}))
-			Expect(err.Error()).To(ContainSubstring(identity.ErrDecodeIdentity.Error()))
-		})
-
-	})
-
-	Context("Interface compatibility", func() {
-		It("should verify XRHID implements IdentityTokenEncapsulator", func() {
-			var encapsulator identity.OrganizationIDProvider
-			xrhid := &identity.XRHID{
-				Identity: identity.Identity{
-					OrgID: "interface-test-org",
-				},
-			}
-
-			encapsulator = xrhid
-			Expect(encapsulator.GetOrganizationID()).To(Equal("interface-test-org"))
-		})
-
-		It("should verify XRHID implements IdentityTokenEncapsulator when the org is not under the identity but under the internal field", func() {
-			var encapsulator identity.OrganizationIDProvider
-			xrhid := &identity.XRHID{
-				Identity: identity.Identity{
-					Internal: identity.Internal{
-						OrgID: "interface-test-org",
-					},
-				},
-			}
-
-			encapsulator = xrhid
-			Expect(encapsulator.GetOrganizationID()).To(Equal("interface-test-org"))
-		})
-
-		It("should verify OAuthEncapsulator implements IdentityTokenEncapsulator", func() {
-			var encapsulator identity.OrganizationIDProvider
-			oauth := identity.OAuthID{}
-
-			encapsulator = oauth
-			Expect(encapsulator.GetOrganizationID()).To(Equal("1"))
-		})
-	})
-})
